+++ spython/main/apps.py
@@ -10,7 +10,8 @@ def apps(self, image=None, full_path=False, root=""):
        return list of SCIF apps in image. The Singularity software serves
        a scientific filesystem integration that will install apps to
        /scif/apps and associated data to /scif/data. For more information 
-       about SCIF, see https://sci-f.github.io
+       about SCIF, see https://sci-f.github.io. Note that this seems
+       to be deprecated in Singularity 3.x.
 
        Parameters
        ==========