+++ spython/oci/cmd/states.py
@@ -8,7 +8,9 @@
 import json
 
 
-def state(self, container_id=None, sudo=None, sync_socket=None):
+def state(
+    self, container_id=None, sudo=None, sync_socket=None, singularity_options=None
+):
 
     """ get the state of an OciImage, if it exists. The optional states that
         can be returned are created, running, stopped or (not existing).
@@ -22,6 +24,7 @@ def state(self, container_id=None, sudo=None, sync_socket=None):
         sudo: Add sudo to the command. If the container was created by root,
               you need sudo to interact and get its state.
         sync_socket: the path to the unix socket for state synchronization
+        singularity_options: a list of options to provide to the singularity client
 
         Returns
         =======
@@ -32,7 +35,7 @@ def state(self, container_id=None, sudo=None, sync_socket=None):
     container_id = self.get_container_id(container_id)
 
     # singularity oci state
-    cmd = self._init_command("state")
+    cmd = self._init_command("state", singularity_options)
 
     if sync_socket is not None:
         cmd = cmd + ["--sync-socket", sync_socket]
@@ -50,7 +53,9 @@ def state(self, container_id=None, sudo=None, sync_socket=None):
             return json.loads(result)
 
 
-def _state_command(self, container_id=None, command="start", sudo=None):
+def _state_command(
+    self, container_id=None, command="start", sudo=None, singularity_options=None
+):
 
     """ A generic state command to wrap pause, resume, kill, etc., where the
         only difference is the command. This function will be unwrapped if the
@@ -63,6 +68,7 @@ def _state_command(self, container_id=None, command="start", sudo=None):
         ==========
         container_id: the id to start.
         command: one of start, resume, pause, kill, defaults to start.
+        singularity_options: a list of options to provide to the singularity client
         sudo: Add sudo to the command. If the container was created by root,
               you need sudo to interact and get its state.
 
@@ -74,7 +80,7 @@ def _state_command(self, container_id=None, command="start", sudo=None):
     container_id = self.get_container_id(container_id)
 
     # singularity oci state
-    cmd = self._init_command(command)
+    cmd = self._init_command(command, singularity_options)
 
     # Finally, add the container_id
     cmd.append(container_id)
@@ -83,7 +89,7 @@ def _state_command(self, container_id=None, command="start", sudo=None):
     return self._run_and_return(cmd, sudo)
 
 
-def start(self, container_id=None, sudo=None):
+def start(self, container_id=None, sudo=None, singularity_options=None):
 
     """ start a previously invoked OciImage, if it exists.
 
@@ -100,10 +106,12 @@ def start(self, container_id=None, sudo=None):
         =======
         return_code: the return code to indicate if the container was started.
     """
-    return self._state_command(container_id, sudo=sudo)
+    return self._state_command(
+        container_id, sudo=sudo, singularity_options=singularity_options
+    )
 
 
-def kill(self, container_id=None, sudo=None, signal=None):
+def kill(self, container_id=None, sudo=None, signal=None, singularity_options=None):
 
     """ stop (kill) a started OciImage container, if it exists
 
@@ -114,6 +122,7 @@ def kill(self, container_id=None, sudo=None, signal=None):
         ==========
         container_id: the id to stop.
         signal: signal sent to the container (default SIGTERM)
+        singularity_options: a list of options to provide to the singularity client
         sudo: Add sudo to the command. If the container was created by root,
               you need sudo to interact and get its state.
 
@@ -125,7 +134,7 @@ def kill(self, container_id=None, sudo=None, signal=None):
     container_id = self.get_container_id(container_id)
 
     # singularity oci state
-    cmd = self._init_command("kill")
+    cmd = self._init_command("kill", singularity_options)
 
     # Finally, add the container_id
     cmd.append(container_id)
@@ -138,7 +147,7 @@ def kill(self, container_id=None, sudo=None, signal=None):
     return self._run_and_return(cmd, sudo)
 
 
-def resume(self, container_id=None, sudo=None):
+def resume(self, container_id=None, sudo=None, singularity_options=None):
     """ resume a stopped OciImage container, if it exists
 
         Equivalent command line example:      
@@ -147,6 +156,7 @@ def resume(self, container_id=None, sudo=None):
         Parameters
         ==========
         container_id: the id to stop.
+        singularity_options: a list of options to provide to the singularity client
         sudo: Add sudo to the command. If the container was created by root,
               you need sudo to interact and get its state.
 
@@ -154,10 +164,15 @@ def resume(self, container_id=None, sudo=None):
         =======
         return_code: the return code to indicate if the container was resumed.
     """
-    return self._state_command(container_id, command="resume", sudo=sudo)
+    return self._state_command(
+        container_id,
+        command="resume",
+        sudo=sudo,
+        singularity_options=singularity_options,
+    )
 
 
-def pause(self, container_id=None, sudo=None):
+def pause(self, container_id=None, sudo=None, singularity_options=None):
     """ pause a running OciImage container, if it exists
 
         Equivalent command line example:      
@@ -166,6 +181,7 @@ def pause(self, container_id=None, sudo=None):
         Parameters
         ==========
         container_id: the id to stop.
+        singularity_options: a list of options to provide to the singularity client
         sudo: Add sudo to the command. If the container was created by root,
               you need sudo to interact and get its state.
 
@@ -173,4 +189,9 @@ def pause(self, container_id=None, sudo=None):
         =======
         return_code: the return code to indicate if the container was paused.
     """
-    return self._state_command(container_id, command="pause", sudo=sudo)
+    return self._state_command(
+        container_id,
+        command="pause",
+        sudo=sudo,
+        singularity_options=singularity_options,
+    )